Modified code related to ORDER and OUTPUT directive so that COPY classes can include multiple groups.
The linker output is based on groups, so a COPY class picks up the group of the
class it is duplicating from. However, initialized far data will span more than one group if it totals > 64K bytes, in which case the linker iterates segments in the class, and any segment not in the same group as the previous one causes all segments in the new group to be iterated.  This is done for group address calculation and loadfile data extraction.


git-svn-id: http://svn.cmeerw.net/openwatcom/trunk@6225 9848d260-d8c2-4c77-ac60-93a84d57cbad