ChangeList: Make the Fortran Run Time expose the saved stack pointer for when crashes occur in the IO subsystems.
The variable is now called FRT_SaveESP and can be accessed in the debugger to aid in getting a call stack.

With the current code, the way to do this is so: when the crash occurs in the IO system, change the ESP value to _FRT_SaveESP and change EBP to _FRT_SaveESP+30 [see RdWrCommon for why 0x30]. Close and reopen the code calls window for an accurate stack call.


git-svn-id: http://svn.cmeerw.net/openwatcom/trunk@8158 9848d260-d8c2-4c77-ac60-93a84d57cbad