ChangeList: The _floodfill() graph library function has been fixed.
Note: The graph lib is abusing the pascal calling convention to
define custom calls. The graph lib broke because of changes to the
calling convention processing in the compiler, but I hesitate to call
that a bug. Ideally the graph lib shouldn't be abusing 'pascal' and
custom calling conventions should be defined in internal headers.


git-svn-id: http://svn.cmeerw.net/openwatcom/trunk@10653 9848d260-d8c2-4c77-ac60-93a84d57cbad